+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inspecto-dev/cli
 
+## 0.3.12
+
+### Patch Changes
+
+- release 0.3.12
+- Updated dependencies
+  - @inspecto-dev/types@0.3.12
+
+## 0.3.11
+
+### Patch Changes
+
+- release 0.3.11
+- Updated dependencies
+  - @inspecto-dev/types@0.3.11
+
 ## 0.3.10
 
 ### Patch Changes